backend/requirements.txt
# Основные Django зависимости (фиксированные версии для стабильности)
Django==4.2.16
requests==2.31.0

# Google API библиотеки
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# Дополнительные библиотеки для работы с Google
PyJWT==2.8.0
cryptography==41.0.8

# Celery и Redis для асинхронных задач
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Временные зоны
pytz==2023.3

# Дополнительные зависимости для работы приложения
Pillow==10.1.0
python-dateutil==2.8.2

# Зависимости для работы с базой данных
sqlite3  # Встроенный в Python

# Telegram интеграция
django-telethon==1.4.0
telethon==1.34.0
django-telethon-session==0.8.1
qrcode[pil]==7.4.2

# Django REST Framework
djangorestframework==3.16.1
django-cors-headers==4.9.0

# Дополнительные зависимости
beautifulsoup4==4.12.2
lxml==4.9.3
